re-organize source code
